refactoring interaction worked
- now to refactor hp...
